// src/app/page/master/golonganPegawai-form/golonganPegawai-form.interface.ts
import {EventEmitter} from '@angular/core';

export interface GolonganPegawai {
    kodeGolonganPegawai?: string;
	namaGolonganPegawai?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	noUrut?: string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	Version?: any;
	label?: any;
	kdDepartemen?: string;
	id_kode?: any;
}